[NAYB-138]refactor: PG API 호출을 트랜잭션에서 분리
