+++ src/test/java/com/prgrms/nabmart/base/BaseControllerTest.java
@@ -12,12 +12,14 @@ import com.prgrms.nabmart.domain.event.service.EventService;
 import com.prgrms.nabmart.domain.item.service.ItemService;
 import com.prgrms.nabmart.domain.item.service.LikeItemService;
 import com.prgrms.nabmart.domain.order.service.OrderService;
+import com.prgrms.nabmart.domain.payment.service.PaymentClient;
 import com.prgrms.nabmart.domain.payment.service.PaymentService;
 import com.prgrms.nabmart.domain.review.service.ReviewService;
 import com.prgrms.nabmart.domain.user.service.UserService;
 import com.prgrms.nabmart.global.auth.oauth.client.OAuthRestClient;
 import com.prgrms.nabmart.global.auth.service.RiderAuthenticationService;
 import com.prgrms.nabmart.global.auth.support.AuthFixture;
+import com.prgrms.nabmart.global.infrastructure.ApiService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -73,6 +75,12 @@ public abstract class BaseControllerTest {
 
     @MockBean
     protected PaymentService paymentService;
+
+    @MockBean
+    protected ApiService apiService;
+
+    @MockBean
+    protected PaymentClient paymentClient;
 
     @MockBean
     protected ReviewService reviewService;
