+++ src/test/java/com/prgrms/nabmart/domain/payment/controller/PaymentControllerTest.java
@@ -26,13 +26,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
 
-@AutoConfigureRestDocs
-@WebMvcTest(PaymentController.class)
 public class PaymentControllerTest extends BaseControllerTest {
 
     @Value("${payment.toss.success-url}")
@@ -164,6 +160,4 @@ public class PaymentControllerTest extends BaseControllerTest {
                 ));
         }
     }
-
-
 }
